requirements.txt
Flask
sqlite3
pycryptodomex
argon2-cffi
```